# [StaticWebAssets] Remove multiple file accesses per asset